This merge was done by isolating the changes that are exclusive to our
v6.1.55/nxp/master branch and porting them directly onto the lf-6.6.3-1.0.0
tag, ignoring any conflicts in files that we haven't customized.

https://onedigi.atlassian.net/browse/DEL-9003

+++ drivers/rpmsg/imx_rpmsg_tty.c
@@ -78,8 +78,8 @@ static void rpmsgtty_close(struct tty_struct *tty, struct file *filp)
 	return tty_port_close(tty->port, tty, filp);
 }
 
-static int rpmsgtty_write(struct tty_struct *tty, const unsigned char *buf,
-			 int total)
+static ssize_t rpmsgtty_write(struct tty_struct *tty, const unsigned char *buf,
+			 size_t total)
 {
 	int count, ret = 0;
 	const unsigned char *tbuf;
